# docs/openapi.yaml
openapi: 3.1.0
info:
  title: php-news API
  version: 1.0.0
  description: |
    HTTP API of the php-news project.

    Two surfaces:
      - `/api/*`        public client API (no auth).
      - `/admin/api/*`  admin API protected by JWT bearer (obtained from `/admin/api/login_check`).
servers:
  - url: http://localhost:8080
...
security: []
paths:
...
  /api/article/{article_id}:
    get:
      tags: [client]
      operationId: getArticle
      summary: Get one article by id.
      parameters:
        - name: article_id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Article found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Article' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # Reusable response wrappers — used by many admin endpoints,
  # kept here to avoid repeating the same body inline 10+ times.
  responses:
...
    NotFound:
      description: Domain entity not found (mapped from `EntityNotFoundException`).
      content:
        application/json:
          schema:
            type: object
            required: [ok, message]
            properties:
              ok:      { type: boolean, enum: [false] }
              message: { type: string }
...
  schemas:
    Article:
      type: object
      required: [id, title, shortDescription, content, numberOfViews, isTop, categories]
      properties:
        id:               { type: integer, minimum: 1 }
        title:            { type: string, minLength: 1, maxLength: 255 }
        shortDescription: { type: string, minLength: 1, maxLength: 500 }
        content:          { type: string, minLength: 1 }
        numberOfViews:    { type: integer, minimum: 0 }
        isTop:            { type: boolean }
        image:
          type: [object, 'null']
          properties:
            id:       { type: [integer, 'null'], minimum: 1 }
            fileName: { type: [string, 'null'] }
        categories:
          type: array
          items: { $ref: '#/components/schemas/CategoryPreview' }
    CategoryPreview:
      type: object
      required: [id, title]
      properties:
        id:    { type: integer, minimum: 1 }
        title: { type: string, minLength: 1, maxLength: 255 }
...